src/requirements.txt
flask
mysql-connector-python==8.0.11